feat(test): Add JSON output to test cmd
