+++ cmd/fossa/test.go
@@ -19,6 +19,10 @@ const pollRequestDelay = time.Duration(8) * time.Second
 const buildsEndpoint = "/api/revisions/%s/build"
 const revisionsEndpoint = "/api/revisions/%s"
 
+type testConfig struct {
+	timeout time.Duration
+}
+
 type buildResponse struct {
 	ID    int
 	Error string
@@ -63,6 +67,8 @@ type revisionResponse struct {
 
 type issueResponse struct {
 	Resolved bool
+	Type     string
+	// RevisionID string // TODO: We need to get this information from /api/issues?fromRevision=
 }
 
 func getRevision(endpoint, apiKey, project, revision string) (revisionResponse, error) {
@@ -91,9 +97,7 @@ func getRevision(endpoint, apiKey, project, revision string) (revisionResponse, 
 	return revisionJSON, nil
 }
 
-func doTest(race chan testResult, endpoint, apiKey, project, revision string) {
-	s := spinner.New(spinner.CharSets[11], 100*time.Millisecond)
-	s.Writer = os.Stderr
+func doTest(s *spinner.Spinner, race chan testResult, endpoint, apiKey, project, revision string) {
 	s.Suffix = " Waiting for analysis to complete..."
 	s.Start()
 
@@ -101,17 +105,14 @@ buildLoop:
 	for {
 		build, err := getBuild(endpoint, apiKey, project, revision)
 		if err != nil {
-			s.Stop()
 			race <- testResult{err: err, issues: nil}
 			return
 		}
 		testLogger.Debugf("Got build: %#v", build)
 		switch build.Task.Status {
 		case "SUCCEEDED":
-			s.Stop()
 			break buildLoop
 		case "FAILED":
-			s.Stop()
 			race <- testResult{err: fmt.Errorf("failed to analyze build #%d: %s (visit FOSSA or contact support@fossa.io)", build.ID, build.Error), issues: nil}
 			return
 		default:
@@ -125,13 +126,11 @@ buildLoop:
 	for {
 		revision, err := getRevision(endpoint, apiKey, project, revision)
 		if err != nil {
-			s.Stop()
 			race <- testResult{err: err, issues: nil}
 			return
 		}
 		testLogger.Debugf("Got revision: %#v", revision)
 		if len(revision.Meta) > 0 && revision.Meta[0].LastScan != "" {
-			s.Stop()
 			race <- testResult{err: nil, issues: revision.Issues}
 			return
 		}
@@ -150,15 +149,43 @@ func testCmd(c *cli.Context) {
 		testLogger.Fatalf("Could not load configuration: %s", err.Error())
 	}
 
+	s := spinner.New(spinner.CharSets[11], 100*time.Millisecond)
+	s.Writer = os.Stderr
 	race := make(chan testResult, 1)
-	go doTest(race, config.endpoint, config.apiKey, config.project, config.revision)
+	go doTest(s, race, config.endpoint, config.apiKey, config.project, config.revision)
 	select {
 	case result := <-race:
+		s.Stop()
 		if result.err != nil {
 			testLogger.Fatalf("Could not execute test: %s", result.err.Error())
 		}
 		testLogger.Debugf("Test succeeded: %#v", result.issues)
-	case <-time.After(config.timeout):
-		testLogger.Fatalf("API requests timed out")
+		if len(result.issues) == 0 {
+			fmt.Fprintln(os.Stderr, "Test passed! 0 issues found")
+		} else {
+			var unresolvedIssues []issueResponse
+			for _, issue := range result.issues {
+				if !issue.Resolved {
+					unresolvedIssues = append(unresolvedIssues, issue)
+				}
+			}
+
+			pluralizedIssues := "issues"
+			if len(unresolvedIssues) == 1 {
+				pluralizedIssues = "issue"
+			}
+			fmt.Fprintf(os.Stderr, "Test failed! %d %s found\n", len(unresolvedIssues), pluralizedIssues)
+
+			issues, err := json.Marshal(unresolvedIssues)
+			if err != nil {
+				testLogger.Fatalf("Could not marshal unresolved issues: %s", err)
+			}
+			fmt.Println(string(issues))
+
+			os.Exit(1)
+		}
+	case <-time.After(config.testConfig.timeout):
+		s.Stop()
+		testLogger.Fatalf("Timed out while waiting for issue report")
 	}
 }
